+++ Assets/Materials/BlueCircle.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9f478c8cfb9e8434da364a3ed57eb2bf
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
