+++ Assets/Scripts/FluidDemo/ParticleVisualization.cs
@@ -12,13 +12,15 @@ namespace FluidSimulation
         public GameObject solidParticlePrefab;
 
         private Dictionary<int, GameObject> _particles;
+        private ParticleSystem _particleSystem; 
 
         private void Awake()
         {
             _particles = new Dictionary<int, GameObject>();
+            _particleSystem = GetComponent<ParticleSystem>();
         }
 
-        public void AddParticle(int id,FluidSubstance substance)
+        public void AddParticle(int id, FluidSubstance substance, Vector2 position)
         {
             if (_particles.ContainsKey(id))
             {
@@ -30,7 +32,17 @@ namespace FluidSimulation
             particle.name = "Particle " + id.ToString();
             _particles.Add(id, particle);
 
+            particle.transform.position = new Vector3(position.x, position.y, 0f);
 
+/*
+            var emitParams = new ParticleSystem.EmitParams
+            {
+                position = new Vector3(position.x, position.y, 0f),
+            };
+
+            _particleSystem.Emit(emitParams, 1);
+*/
+            
             GameObject PrefabFor(FluidSubstance substance) => substance switch
             {
                 FluidSubstance.SomeLiquid => liquidParticlePrefab,
@@ -38,7 +50,8 @@ namespace FluidSimulation
                 FluidSubstance.SomeSolid => solidParticlePrefab,
                 _ => throw new ArgumentOutOfRangeException(nameof(substance), substance, null)
             };
-
+            
+           
         }
 
         public void RemoveParticle(int id)
@@ -74,11 +87,25 @@ namespace FluidSimulation
 
             var particle = _particles.GetValueOrDefault(id);
             particle.transform.position = new Vector3(position.x, position.y, 0f);
+
+
+/*
+            Particle[] particles = particleEmitter.particles;
+
+            // Do changes
+            for (int i = 0; i < particles.Length; i++)
+            {
+                particles<em>.position = Vector3.MoveTowards(particleEmitter.particles_.position,
+                    spot.transform.position, Time.deltaTime * speed);
+                _ </em >
+            }
+            
+            */
         }
 
         public void ColorParticle(int id, Color color)
-            => _particles[id].GetComponentInChildren<SpriteRenderer>().color = color;
-        
+//            => _particles[id].GetComponentInChildren<SpriteRenderer>().color = color;
+        {}      
         
         public Color RandomColor => new Color(UnityEngine.Random.value, UnityEngine.Random.value, UnityEngine.Random.value);
     }
