+++ Assets/Scripts/FluidSimulation/Fluid.cs
@@ -1,0 +1,8 @@
+namespace FluidSimulation
+{
+    public abstract record Fluid(string Name, float Density);
+
+    public record Liquid(string Name, float Density, float Viscosity) : Fluid(Name, Density);
+    public record Gas(string Name, float Density, float Viscosity)    : Fluid(Name, Density);
+    public record Solid(string Name, float Density)                   : Fluid(Name, Density);  // Non-fluid fluid :)
+}
