+++ Assets/SmokeParticle.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 07658e9a2ca554275a154b74025271b1
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
